just converted another 2 api calls from simple to mvvm style coding

+++ frontend/lib/App2/MVVM/Model/products.dart
@@ -1,10 +1,11 @@
 class Products {
+  
   String? sId;
   String? userId;
   String? name;
   String? shortDescription;
   String? description;
-  double? price;
+  int? price;
   String? image;
   String? createdAt;
   String? updatedAt;
